+++ NEWS.md
@@ -1,5 +1,14 @@
 # odataR
 
+## odataR 0.1.2
+
+* 08nov2019   
+   + odataR_get_table argument `decode` added and  
+   + odataR_get_table can handle `$count`
+   + odataR_get_table and odataR_get_meta cleanup for determining dimensions
+   + odataR_query argument `debug` added (also in higher level functions)
+   + reorganised code: handling of query moved to odataR_query
+
 ## odataR 0.1.1
 
 * 05nov2019